将int和double基本数据类型都改成了Integer和Double封装类，为了适应Bmob云端数据库的需要

+++ src/com/libra/stockanalysisi/bean/Stock.java
@@ -8,87 +8,87 @@ public class Stock extends BaseStock {
 	 * 
 	 */
 	private static final long serialVersionUID = 4173556025373242718L;
-	private double m_TodayStartPri; /* 今日开盘价 */
-	private double m_YestodEndPri; /* 昨日收盘价 */
-	private double m_NowPri; /* 当前价格 */
-	private double m_TodayMax; /* 今日最高价 */
-	private double m_TodayMin; /* 今日最低价 */
-	private double m_CompetitivePri; /* 竞买价 */
-	private double m_ReservePri; /* 竞卖价 */
-	private double m_TraNumber; /* 成交量 */
-	private double m_TraAmount; /* 成交金额 */
+	private Double m_TodayStartPri; /* 今日开盘价 */
+	private Double m_YestodEndPri; /* 昨日收盘价 */
+	private Double m_NowPri; /* 当前价格 */
+	private Double m_TodayMax; /* 今日最高价 */
+	private Double m_TodayMin; /* 今日最低价 */
+	private Double m_CompetitivePri; /* 竞买价 */
+	private Double m_ReservePri; /* 竞卖价 */
+	private Integer m_TraNumber; /* 成交量 */
+	private Integer m_TraAmount; /* 成交金额 */
 	// 2012-12-11
 	private Date m_Date; /* 日期 */
 
-	public double getTodayStartPri() {
+	public Double getTodayStartPri() {
 		return m_TodayStartPri;
 	}
 
-	public void setTodayStartPri(double pTodayStartPri) {
+	public void setTodayStartPri(Double pTodayStartPri) {
 		this.m_TodayStartPri = pTodayStartPri;
 	}
 
-	public double getYestodEndPri() {
+	public Double getYestodEndPri() {
 		return m_YestodEndPri;
 	}
 
-	public void setYestodEndPri(double pYestodEndPri) {
+	public void setYestodEndPri(Double pYestodEndPri) {
 		this.m_YestodEndPri = pYestodEndPri;
 	}
 
-	public double getNowPri() {
+	public Double getNowPri() {
 		return m_NowPri;
 	}
 
-	public void setNowPri(double pNowPri) {
+	public void setNowPri(Double pNowPri) {
 		this.m_NowPri = pNowPri;
 	}
 
-	public double getTodayMax() {
+	public Double getTodayMax() {
 		return m_TodayMax;
 	}
 
-	public void setTodayMax(double pTodayMax) {
+	public void setTodayMax(Double pTodayMax) {
 		this.m_TodayMax = pTodayMax;
 	}
 
-	public double getTodayMin() {
+	public Double getTodayMin() {
 		return m_TodayMin;
 	}
 
-	public void setTodayMin(double pTodayMin) {
+	public void setTodayMin(Double pTodayMin) {
 		this.m_TodayMin = pTodayMin;
 	}
 
-	public double getCompetitivePri() {
+	public Double getCompetitivePri() {
 		return m_CompetitivePri;
 	}
 
-	public void setCompetitivePri(double pCompetitivePri) {
+	public void setCompetitivePri(Double pCompetitivePri) {
 		this.m_CompetitivePri = pCompetitivePri;
 	}
 
-	public double getReservePri() {
+	public Double getReservePri() {
 		return m_ReservePri;
 	}
 
-	public void setReservePri(double pReservePri) {
+	public void setReservePri(Double pReservePri) {
 		this.m_ReservePri = pReservePri;
 	}
 
-	public double getTraNumber() {
+	public Integer getTraNumber() {
 		return m_TraNumber;
 	}
 
-	public void setTraNumber(double pTraNumber) {
+	public void setTraNumber(Integer pTraNumber) {
 		this.m_TraNumber = pTraNumber;
 	}
 
-	public double getTraAmount() {
+	public Integer getTraAmount() {
 		return m_TraAmount;
 	}
 
-	public void setTraAmount(double pTraAmount) {
+	public void setTraAmount(Integer pTraAmount) {
 		this.m_TraAmount = pTraAmount;
 	}
 
@@ -100,10 +100,10 @@ public class Stock extends BaseStock {
 		this.m_Date = pDate;
 	}
 
-	public Stock(String m_Gid, String m_Name, double m_TodayStartPri,
-			double m_YestodEndPri, double m_NowPri, double m_TodayMax,
-			double m_TodayMin, double m_CompetitivePri, double m_ReservePri,
-			double m_TraNumber, double m_TraAmount, Date m_Date) {
+	public Stock(String m_Gid, String m_Name, Double m_TodayStartPri,
+			Double m_YestodEndPri, Double m_NowPri, Double m_TodayMax,
+			Double m_TodayMin, Double m_CompetitivePri, Double m_ReservePri,
+			Integer m_TraNumber, Integer m_TraAmount, Date m_Date) {
 		super();
 		this.m_Gid = m_Gid;
 		this.m_Name = m_Name;
